Updated when the the javascript function runs.
Why:
I found when implementing the gem that on a page refresh the dependent fields were applied. However when the route to the new form was followed the dependency was not functioning.

This PR:
Adds to the code instructions that on every page load that the function should occur. Adding this to my project means that whether the page is refreshed or loaded having followed its route the dependent fields display correctly.