Updated the README
Why:
On further understanding of the issue I realize that my initial commit was not completely accurate.

This commit:
Includes the original jQuery and my updated jQuery. The original code will work if that app does not have Turbolinks included. If Turbolinks is active in the app my updated jQuery is necessary for the function to run on every page load.